Valida lançamentos no iDiário no cadastro de turmas

+++ ieducar/modules/ComponenteCurricular/Model/TurmaDataMapper.php
@@ -1,5 +1,9 @@
 <?php
 
+use App\Exceptions\SchoolClass\HasDataInDiario;
+use App\Models\LegacyDiscipline;
+use App\Services\iDiarioService;
+
 require_once 'CoreExt/DataMapper.php';
 require_once 'ComponenteCurricular/Model/Turma.php';
 
@@ -56,14 +60,14 @@ class ComponenteCurricular_Model_TurmaDataMapper extends CoreExt_DataMapper
      *
      *
      *
-     * @param int   $anoEscolar  O código do ano escolar/série.
-     * @param int   $escola      O código da escola.
-     * @param int   $turma       O código da turma.
+     * @param int $anoEscolar O código do ano escolar/série.
+     * @param int $escola O código da escola.
+     * @param int $turma O código da turma.
      * @param array $componentes (id => integer, cargaHoraria => float|null)
-     *
-     * @throws Exception
+     * @param iDiarioService|null $iDiarioService
+     * @throws CoreExt_DataMapper_Exception
      */
-    public function bulkUpdate($anoEscolar, $escola, $turma, array $componentes)
+    public function bulkUpdate($anoEscolar, $escola, $turma, array $componentes, $iDiarioService = null)
     {
         $update = $insert = $delete = [];
 
@@ -101,8 +105,19 @@ class ComponenteCurricular_Model_TurmaDataMapper extends CoreExt_DataMapper
 
         $delete = array_diff(array_keys($objects), array_keys($insert));
 
+        $erros = [];
         foreach ($delete as $id) {
+            if ($iDiarioService && $iDiarioService->getClassroomsActivityByDiscipline([$turma], $id)) {
+                $discipline = LegacyDiscipline::find($id);
+                $erros[] = sprintf('Não é possível desvincular "%s" pois já existem notas, faltas e/ou pareceres lançados para este componente nesta turma no iDiário.', $discipline->nome);
+                continue;
+            }
+
             $this->delete($objects[$id]);
+        }
+
+        if ($erros) {
+            throw new HasDataInDiario($erros);
         }
 
         foreach ($insert as $entry) {
